Create Predictive_Parser.c
Implementing a predictive parser in C for a given expression grammar allows you to parse and evaluate arithmetic expressions according to the specified rules. Here's the implementation of a predictive parser for the given grammar:

This C program takes an arithmetic expression as input and uses a series of functions to parse it according to the provided grammar rules. If the expression is valid, it prints "Valid expression!"; otherwise, it indicates that the expression is invalid. The program also handles parentheses and digits as part of the grammar rules.